requirements_no_pyarrow.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
mlxtend>=0.22
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
streamlit>=1.35
sqlalchemy>=2.0
watchdog>=3.0
pyyaml>=6.0
reportlab>=4.0
requests>=2.31
pytest>=7.4
pytest-cov>=4.1
httpx>=0.27
black>=23.7
flake8>=6.0
pyinstaller>=6.0
